feat(model): ✨ 完成model功能
